requirements.txt
# --------- pytorch --------- #
lightning==2.5.0.post0
pytorch-lightning==2.5.0.post0
torch==2.5.1
torchvision==0.20.1


# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
tensorboard==2.20.0

# --------- linters --------- #
pre-commit==4.0.1      # hooks for applying linters on commit

# --------- others --------- #
rich==13.9.4
rasterio==1.4.3
fiona==1.10.1
pyarrow==18.1.0
geopandas==1.0.1
numpy==2.2.1
scipy==1.15.0
seaborn==0.13.2
scikit-image==0.25.0
openpyxl==3.1.5
timm==1.0.15
joblib==1.5.1
pystac-client==0.9.0
teledetection==0.1.7
geedim==1.9.1
scikit-learn==1.6.0
geefetch @ git+https://github.com/ThomasBoudras/geefetch_sera_h.git@main